Injected parent classes to ReferenceBindingSave and ValueReference. The one for ReferenceBindingSave handles deferred saving, for the purpose of hooking in a different custom serializer, and the ValueReference is a generic parent class that allows listening on changes without taking in the value as an argument.
