+++ Runtime/Scripts/Ratferences/ReferenceBindingSave.cs
@@ -5,10 +5,10 @@ using System.Collections.Generic;
 using UnityEngine;
 
 namespace Ratferences {
-	/// <summary>
-	/// Interface that manages saving the actual options. This is normally the struct that holds the save information.
-	/// </summary>
-	public interface ISavable {
+    /// <summary>
+    /// Interface that manages saving the actual options. This is normally the struct that holds the save information.
+    /// </summary>
+    public interface ISavable {
 		void Save();
 	}
 
@@ -16,8 +16,7 @@ namespace Ratferences {
 	/// Class that manages the binding of value references to some saveable state.
 	/// </summary>
 	/// <typeparam name="R"></typeparam>
-	public abstract class ReferenceBindingSave<R> : MonoBehaviour where R : ISavable {
-
+	public abstract class ReferenceBindingSave<R> : SaveDelayer where R : ISavable {
 		public enum SaveType {
 			Immediate,
 			AfterDelay,
@@ -37,23 +36,11 @@ namespace Ratferences {
 			}
 		}
 
-		[Tooltip("How long after the last reference change to wait until saving the game. Useful if options are changed a lot in a small amount of time, like a user messing with a slider.")]
-		public float SaveDelay = 1f;
-		[Tooltip("Maximum time to wait to save the game. Useful if you're constantly changing some scriptable objects and want to minimize the worst-case save delay scenario. NOTE: In this case, you should probably be disabling triggersSave on the binding.")]
-		public float MaxSaveDelay = 15f;
-
 		private readonly Dictionary<FloatReference, Binding<float, FloatReference, R>> _floatBindings = new Dictionary<FloatReference, Binding<float, FloatReference, R>>();
 		private readonly Dictionary<BoolReference, Binding<bool, BoolReference, R>> _boolBindings = new Dictionary<BoolReference, Binding<bool, BoolReference, R>>();
 		private readonly Dictionary<IntReference, Binding<int, IntReference, R>> _intBindings = new Dictionary<IntReference, Binding<int, IntReference, R>>();
 		private readonly Dictionary<StringReference, Binding<string, StringReference, R>> _stringBindings = new Dictionary<StringReference, Binding<string, StringReference, R>>();
 		private readonly Dictionary<Vector3Reference, Binding<Vector3, Vector3Reference, R>> _vector3Bindings = new Dictionary<Vector3Reference, Binding<Vector3, Vector3Reference, R>>();
-		private float _delayableSaveTime;
-		private float _undelayableSaveTime;
-		private Coroutine _saveCoroutine = null;
-		// Theoretically this variable is redundant because I *should* be able
-		// to check if (_saveCoroutine != null), but a null coroutine was still
-		// going into the block. Probably Unity doing its weird fake null thing.
-		private bool _isCoroutineActive = false;
 
 		/// <summary>
 		/// Loads the initial state, creating a new instance if necessary.
@@ -130,30 +117,7 @@ namespace Ratferences {
 			AdditionalUnbindings();
 		}
 
-		public void ScheduleSave() {
-			_delayableSaveTime = Time.unscaledTime + SaveDelay;
-			_undelayableSaveTime = Mathf.Min(Time.unscaledTime + MaxSaveDelay, _undelayableSaveTime);
-			if (!_isCoroutineActive) {
-				_isCoroutineActive = true;
-				_saveCoroutine = StartCoroutine(SaveOptionsAfterDelay());
-			}
-		}
-
-		IEnumerator SaveOptionsAfterDelay() {
-			while (Time.unscaledTime < _delayableSaveTime && Time.unscaledTime < _undelayableSaveTime) {
-				yield return null;
-			}
-			SaveNow();
-		}
-
-		public void SaveNow() {
-			if (_isCoroutineActive) {
-				// Theoretically these should be in alignment, but just in case.
-				if (_saveCoroutine != null) StopCoroutine(_saveCoroutine);
-				_saveCoroutine = null;
-				_isCoroutineActive = false;
-			}
-			_undelayableSaveTime = float.MaxValue;
+		protected override void PerformSave() {
 			GetSave().Save();
 		}
