+++ Runtime/Scripts/Ratferences/SaveType.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace Ratferences {
+    public enum SaveType {
+        Immediate,
+        AfterDelay,
+        DoesntTrigger
+    }
+}
