+++ Runtime/Scripts/Ratferences/SaveType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d33a2e43964fc2440a465683a592a913
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
